some W4 level warnings fixed; some x64 warnings fixed;

+++ src/engine/core/string.h
@@ -36,9 +36,9 @@ LUMIX_ENGINE_API bool startsWith(const char* str, const char* prefix);
 LUMIX_ENGINE_API int stringLength(const char* str);
 LUMIX_ENGINE_API int compareString(const char* lhs, const char* rhs);
 LUMIX_ENGINE_API int compareStringN(const char* lhs, const char* rhs, int length);
-LUMIX_ENGINE_API void copyMemory(void* dest, const void* src, int count);
-LUMIX_ENGINE_API void moveMemory(void* dest, const void* src, int count);
-LUMIX_ENGINE_API void setMemory(void* ptr, uint8 value, int num);
+LUMIX_ENGINE_API void copyMemory(void* dest, const void* src, size_t count);
+LUMIX_ENGINE_API void moveMemory(void* dest, const void* src, size_t count);
+LUMIX_ENGINE_API void setMemory(void* ptr, uint8 value, size_t num);
 LUMIX_ENGINE_API const char* findSubstring(const char* str, const char* substr);
